requirements.txt
asgiref
attrs
autobahn
Automat
cffi
channels
channels_redis
constantly
cryptography
daphne
dj-database-url
Django
django-environ
hyperlink
idna
incremental
msgpack
pillow
psycopg2-binary
pyasn1
pyasn1_modules
pycparser
pyOpenSSL
redis
service-identity
shortuuid
sqlparse
Twisted
txaio
typing_extensions
tzdata
whitenoise
zope.interface